refactor: reorganize imports to use models instead of _types for better clarity
