+++ launcher_core/__init__.py
@@ -5,15 +5,10 @@
 __version__ = "0.3.1"
 
 from .logging_utils import logger
-from .pydantic_models import (
-    Credential,
-    AzureApplication,
-)
 from .check_version import check_version
 from .config.load_launcher_config import ConfigManager, LauncherConfig
-
 # 導入 Pydantic 模型
-from .pydantic_models import *
+from .models.auth import MinecraftUUID, Credential, AzureApplication
 
 from . import (
     command,
@@ -30,7 +25,6 @@ from . import (
     exceptions,
     microsoft_types,
     config,
-    pydantic_models,
 )
 from .utils import sync
 from .mojang import verify_mojang_jwt
@@ -48,36 +42,15 @@ __all__ = [
     "runtime",
     "mrpack",
     "exceptions",
-    "_types",
+    "models",
     "microsoft_types",
     "config",
-    "pydantic_models",
     "logger",
     "sync",
     "Credential",
     "AzureApplication",
     "ConfigManager",
     "LauncherConfig",
-    # Pydantic 模型
-    "MinecraftOptions",
-    "Credential",
-    "LaunchProfile",
-    "ServerInfo",
-    "ModInfo",
-    "LauncherSettings",
-    "JavaInformation",
-    "DownloadInfo",
-    "LibraryInfo",
-    "AssetInfo",
-    "LatestMinecraftVersions",
-    "MinecraftVersionInfo",
-    "FabricMinecraftVersion",
-    "FabricLoader",
-    "QuiltMinecraftVersion",
-    "QuiltLoader",
-    "VanillaLauncherProfile",
-    "VanillaLauncherProfileResolution",
-    "MrpackInformation",
     "MinecraftUUID",
     "__version__",
     "check_version",
