+++ launcher_core/command.py
@@ -16,8 +16,8 @@ from ._internal_types.shared_types import ClientJson, ClientJsonArgumentRule
 from .runtime import get_executable_path
 from .exceptions import VersionNotFound
 from .utils import get_library_version
-from ._types import MinecraftOptions
-from ._types import Credential as AuthCredential
+from .models import MinecraftOptions
+from .models import Credential as AuthCredential
 from .natives import get_natives
 
 __all__ = ["get_minecraft_command"]
