+++ launcher_core/quilt.py
@@ -19,7 +19,7 @@ from ._helper import (
     SUBPROCESS_STARTUP_INFO,
 )
 from .exceptions import VersionNotFound, UnsupportedVersion, ExternalProgramError
-from ._types import QuiltMinecraftVersion, QuiltLoader, CallbackDict
+from .models import QuiltMinecraftVersion, QuiltLoader, CallbackDict
 from .install import install_minecraft_version
 from .utils import is_version_valid
